% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(language, pool)
}
\arguments{
\item{language}{String of language for application ("en" or "fr")}

\item{pool}{String of path of yaml file database configuration (ex. "~/SI_SNOT/dbconfProd.yaml")}
}
\description{
Run the Shiny Application
}
